chore: delete legacy student management components and consolidate architecture
- **Cleanup**: Removed `AddStudentDialog` and its associated sub-components (`AddStudentDialogHeader`, `AddStudentDialogFooter`, `AddStudentFormFields`) in favor of the new `StudentFormDrawer`.
- **Refactor**: Deleted `StudentManageToolbar` and `StudentRowActions`, consolidating their functionality into the revamped `StudentsDataTable` header and row definitions.
- **Organization**: Relocated `StudentsDataTable` and its internal components from the generic `dataTable` folder to a dedicated feature-based directory.
- **Dependency Management**: Updated all imports in `ManageStudentClient` to point to the new, streamlined component locations.
- **Logic Removal**: Purged legacy hooks and constants that were exclusively used by the old dialog system.
