+++ features/manage-list/components/StudentsDataTable/index.tsx
